Make first two functions
The first function, which is capitalize, takes a String as input and
capitalize the first letter, returning the same string with that change.

On the other hand, the second function is reverseString, which returns
the parameter entered reversed.

In the .eslintrc.json document the respective code was added to prevent
the linter from highliting the jest matchers as error.